requirements.txt
pandas
numpy
scikit-learn
nltk
datasets
matplotlib
seaborn
hf_xet
streamlit
PyPDF2
python-docx
joblib